[REF] Acsoo Addon Toupdate: move the command in separate file refs #16
